bugfix in clone_* funcs

+++ std_blocks/ptrig/ptrig.c
@@ -188,7 +188,7 @@ static void ptrig_handle_config(ubx_block_t *b)
 		ERR("failed to set sched_policy.sched_priority to %d", sched_param.sched_priority);
 
 	DBG("%s config: period=%lus:%luus, policy=%s, prio=%d, stacksize=%lu (0=default size)",
-	    b->name, inf->period.sec, inf->period.usec, schedpol_str, sched_param.sched_priority, stacksize);
+	    b->name, inf->period.sec, inf->period.usec, schedpol_str, sched_param.sched_priority, (unsigned long) stacksize);
 }
 
 /* init */
